requirements.txt
flask
sklearn
gunicorn
pillow
numpy
pandas
matplotlib.pyplot